EID-811: Add support for additional eIDAS signing algorithms
The eIDAS SAML Crypto specification allows only two cryptographic
algorithms to be used for signing SAML messages and metadata. One of
these (ECDSA) is already in the default OpenSAML
SignatureValidationConfiguration whitelist and just needed to be
whitelisted in `PKIXSignatureValidationFilterProvider` and
`SignatureValidator`. The other algorithm (RSASSA-PSS) is not in the
default OpenSAML whitelist and isn't supported natively by the Oracle
security providers.

The following changes are included:

- Add the BouncyCastle JCE Provider to the list of Java security
providers.
- Add an `EidasSecurityConfigurationInitializer` which uses the OpenSAML
`Initializer` interface and `InitializationService` mechanism to set
things up.
- Whitelist the RSASSA-PSS (sha256-rsa-MGF1) algorithm in
`PKIXSignatureValidationFilterProvider` and `SignatureValidator`.

Author: @vixus0