EID-1072: Throw exception instead of error if expired cert
The MSA won't start if the trust anchor has an expired cert in it. This
is because the `EidasMetadataResolverRepository` throws an `Error` if
it finds an expired cert. This brings the app down.

This updates so that an Exception is thrown instead which is cauht and
handled, whilst logging. Sensu will pick up any expired certs.